+++ src/main/java/com/example/notas/repository/ProfesoresRepository.java
@@ -2,8 +2,10 @@ package com.example.notas.repository;
 
 import com.example.notas.model.Profesores;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface ProfesoresRepository extends JpaRepository<Profesores, Long> {
+public interface ProfesoresRepository extends JpaRepository<Profesores, Integer> {
     Optional<Profesores> findByEmail(String email);
 }
+
